requirements.txt
beautifulsoup4==4.13.4
edge_tts==7.0.2
openai~=1.88.0
pocketflow==0.0.2
python-dotenv==1.1.0
requests==2.32.4
playwright~=1.53.0
